requirements.txt
pyserial
digi-xbee
libscrc